gdb: restructure pretty printers into different files
